Optimize code

+++ add/add.go
@@ -15,9 +15,12 @@ import (
 	"strings"
 )
 
+const API_URL = "https://p33dswbrne.execute-api.eu-central-1.amazonaws.com/develop/template"
+const SKELETON_FILE = "Skeletonfile.json"
+
 var regex = regexp.MustCompile("\\${(.*?)}")
 
-type projectDescriptor struct {
+type skeletonfile struct {
 	Version     string      `json:"version"`
 	Name        string      `json:"name"`
 	Description string      `json:"description"`
@@ -35,25 +38,28 @@ type response struct {
 	FileUrls []string `json:"fileUrls"`
 }
 
+type request struct {
+	GroupId string `json:"groupId"`
+	Name    string `json:"name"`
+	Version string `json:"version"`
+}
+
 func Add(args []string) {
-	response := getTemplate([]byte(` { "groupId": "fseemann", "name": "maven-domain-module", "version": "latest" } `))
-	files := createTmpFiles(response.FileUrls)
+	requestBody := createRequestBody(args)
 
-	filecontents, err := ioutil.ReadFile(files[0].Name())
-	if err != nil {
-		log.Fatal(err)
-	}
+	response := fetchTemplate(requestBody)
+	templateFiles := createTmpFiles(response.FileUrls)
+	defer removeFiles(templateFiles)
 
-	var pd projectDescriptor
-	if err := json.Unmarshal(filecontents, &pd); err != nil {
-		log.Fatal(err)
-	}
+	skeletonfile := parseSkeletonfile(templateFiles[SKELETON_FILE].Name())
+	variables := readVariables(skeletonfile.Variables)
+	execute(skeletonfile, variables, templateFiles)
+}
 
-	templateFiles := files[1:]
-	variables := readVariables(pd.Variables)
-	for i, struc := range pd.Structure {
+func execute(sf skeletonfile, variables map[string]string, templateFiles map[string]*os.File) {
+	for _, struc := range sf.Structure {
 		actualDir := replaceVariables(struc.Dir, variables)
-		t := template.Must(template.ParseFiles(templateFiles[i].Name()))
+		t := template.Must(template.ParseFiles(templateFiles[struc.Template].Name()))
 		if err := os.MkdirAll(actualDir, os.ModePerm); err != nil {
 			log.Fatal(err)
 		}
@@ -69,19 +75,58 @@ func Add(args []string) {
 		_ = writer.Flush()
 		_ = file.Close()
 	}
-
-	removeFiles(files)
 }
 
-func createTmpFiles(fileUrls []string) []*os.File {
-	files := make([]*os.File, len(fileUrls))
-	for i, url := range fileUrls {
+func parseSkeletonfile(fileName string) skeletonfile {
+	skeletonfileContents, err := ioutil.ReadFile(fileName)
+	if err != nil {
+		log.Fatal(err)
+	}
+	var pd skeletonfile
+	if err := json.Unmarshal(skeletonfileContents, &pd); err != nil {
+		log.Fatal(err)
+	}
+	return pd
+}
+
+func createRequestBody(args []string) []byte {
+	if len(args) != 1 {
+		log.Fatalf("Provide a template identifier")
+	}
+	arg := args[0]
+	splitVersion := strings.Split(arg, ":")
+
+	version := "latest"
+	if len(splitVersion) == 2 {
+		version = splitVersion[1]
+	}
+
+	groupIdAndName := strings.Split(splitVersion[0], "/")
+	if len(groupIdAndName) != 2 {
+		log.Fatalf("Invalid template name.")
+	}
+
+	marshal, err := json.Marshal(request{
+		GroupId: groupIdAndName[0],
+		Name:    groupIdAndName[1],
+		Version: version,
+	})
+	if err != nil {
+		log.Fatalf("Could not create request body.", err)
+	}
+
+	return marshal
+}
+
+func createTmpFiles(fileUrls []string) map[string]*os.File {
+	files := make(map[string]*os.File, len(fileUrls))
+	for _, url := range fileUrls {
 		resp, err := http.Get(url)
 		if err != nil {
 			log.Fatal(err)
 		}
 
-		tmpFile, err := ioutil.TempFile(os.TempDir(), "prefix-")
+		tmpFile, err := ioutil.TempFile(os.TempDir(), "skeleton-template-")
 		if err != nil {
 			removeFiles(files)
 			log.Fatal("Cannot create temporary file", err)
@@ -94,21 +139,31 @@ func createTmpFiles(fileUrls []string) []*os.File {
 		}
 
 		resp.Body.Close()
-		files[i] = tmpFile
+
+		lastIndex := strings.LastIndex(url, "/")
+		filename := url[lastIndex+1:]
+		files[filename] = tmpFile
 	}
 
 	return files
 }
 
-func removeFiles(files []*os.File) {
+func removeFiles(files map[string]*os.File) {
 	for _, f := range files {
-		_ = f.Close()
-		_ = os.Remove(f.Name())
+		removeFile(f)
 	}
 }
 
-func getTemplate(body []byte) response {
-	resp, err := http.Post("https://p33dswbrne.execute-api.eu-central-1.amazonaws.com/develop/template", "application/json", bytes.NewBuffer(body))
+func removeFile(f *os.File) {
+	if f == nil {
+		return
+	}
+	_ = f.Close()
+	_ = os.Remove(f.Name())
+}
+
+func fetchTemplate(body []byte) response {
+	resp, err := http.Post(API_URL, "application/json", bytes.NewBuffer(body))
 	if err != nil {
 		log.Fatal(err)
 	}
